+++ src/Device/DeviceSelector/DeviceList/MoreDeviceInfo.tsx
@@ -10,56 +10,52 @@ import { SerialPort } from '../../../../nrfutil/device/common';
 import { displayedDeviceName } from '../../deviceInfo/deviceInfo';
 import { Device } from '../../deviceSlice';
 
-import './more-device-info.scss';
-
 const Row = ({ children }: { children: ReactNode }) => (
-    <div className="info-row">
-        <div className="flex-space" />
+    <div className=" tw-flex tw-flex-row tw-pr-5">
+        <div className=" tw-w-[68px]" />
         {children}
     </div>
 );
 
 const PcaNumber = ({ device }: { device: Device }) => {
-    if (!device.jlink?.boardVersion) {
-        return null;
-    }
-
-    return <div>{device.jlink.boardVersion}</div>;
-};
-
-const MaybeDeviceName = ({ device }: { device: Device }) => {
-    const hasNickname = device.nickname !== '';
-    if (!hasNickname) {
+    if (!device.devkit?.boardVersion) {
         return null;
     }
 
     return (
-        <div className="name">
-            {displayedDeviceName(device, { respectNickname: false })}
-        </div>
+        <Row>
+            <div>{device.devkit.boardVersion}</div>
+        </Row>
     );
 };
 
-const Serialports = ({ ports }: { ports: SerialPort[] }) => (
-    <ul className="ports">
-        {ports.map(port => (
-            <li key={port.path}>{port.comName}</li>
-        ))}
-    </ul>
-);
+const MaybeDeviceName = ({ device }: { device: Device }) => {
+    if (!device.nickname) {
+        return null;
+    }
+
+    return (
+        <Row>
+            <div>{displayedDeviceName(device, { respectNickname: false })}</div>
+        </Row>
+    );
+};
+
+const Serialports = ({ ports }: { ports: SerialPort[] }) =>
+    ports.length > 0 ? (
+        <Row>
+            <ul className="tw-text-center tw-underline">
+                {ports.map(port => (
+                    <li key={port.path}>{port.comName}</li>
+                ))}
+            </ul>
+        </Row>
+    ) : null;
 
 export default ({ device }: { device: Device }) => (
-    <div className="more-infos">
-        <Row>
-            <PcaNumber device={device} />
-        </Row>
-        <Row>
-            <MaybeDeviceName device={device} />
-        </Row>
-        {device.serialPorts && (
-            <Row>
-                <Serialports ports={device.serialPorts} />
-            </Row>
-        )}
+    <div className="tw-preflight tw-flex tw-flex-col tw-gap-2">
+        <PcaNumber device={device} />
+        <MaybeDeviceName device={device} />
+        <Serialports ports={device.serialPorts ?? []} />
     </div>
 );