+++ src/Device/DeviceSelector/DeviceSelector.tsx
@@ -4,10 +4,12 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
+import { NrfutilDeviceLib } from '../../../nrfutil';
 import { DeviceTraits } from '../../../nrfutil/device/common';
+import logger from '../../logging';
 import useHotKey from '../../utils/useHotKey';
 import {
     clearWaitForDevice,
@@ -22,7 +24,8 @@ import {
     Device,
     deviceIsSelected as deviceIsSelectedSelector,
     selectDevice,
-    selectedSerialNumber,
+    selectedDevice,
+    setSelectedDeviceInfo,
 } from '../deviceSlice';
 import DeviceList from './DeviceList/DeviceList';
 import SelectDevice from './SelectDevice';
@@ -54,7 +57,7 @@ export default ({
     const [deviceListVisible, setDeviceListVisible] = useState(false);
 
     const deviceIsSelected = useSelector(deviceIsSelectedSelector);
-    const selectedSN = useSelector(selectedSerialNumber);
+    const currentDevice = useSelector(selectedDevice);
     const waitingToAutoReconnect = useSelector(getWaitingToAutoReselect);
     const showSelectedDevice = deviceIsSelected || waitingToAutoReconnect;
 
@@ -65,25 +68,47 @@ export default ({
         dispatch(deselectDevice());
     }, [dispatch, onDeviceDeselected]);
 
+    const abortController = useRef<AbortController>();
+
     // Ensure that useCallback is
     // not updated frequently as this
     // will have a side effect to stop and start the hotplug events
     const doSelectDevice = useCallback(
-        (device: Device, autoReselected: boolean) => {
+        async (device: Device, autoReselected: boolean) => {
             dispatch(clearWaitForDevice());
             setDeviceListVisible(false);
             dispatch(selectDevice(device));
             dispatch(setAutoSelectDevice(device));
             onDeviceSelected(device, autoReselected);
+
+            abortController.current?.abort();
+            abortController.current = new AbortController();
+            const deviceInfo = await NrfutilDeviceLib.deviceInfo(
+                device,
+                undefined,
+                undefined,
+                abortController.current
+            );
+            abortController.current = undefined;
+            setSelectedDeviceInfo(deviceInfo);
+
             if (deviceSetupConfig) {
-                dispatch(
-                    setupDevice(
-                        device,
-                        deviceSetupConfig,
-                        onDeviceIsReady,
-                        doDeselectDevice
-                    )
-                );
+                if (device.serialNumber) {
+                    dispatch(
+                        setupDevice(
+                            device,
+                            deviceSetupConfig,
+                            onDeviceIsReady,
+                            doDeselectDevice,
+                            deviceInfo
+                        )
+                    );
+                } else {
+                    logger.warn(
+                        `Selected device has no serial number. Device setup is not supported`
+                    );
+                    onDeviceIsReady(device);
+                }
             }
         },
         [
@@ -145,7 +170,7 @@ export default ({
             <DeviceList
                 isVisible={deviceListVisible}
                 doSelectDevice={(device, autoReselected) => {
-                    if (device.serialNumber === selectedSN) {
+                    if (device.id === currentDevice?.id) {
                         setDeviceListVisible(false);
                         return;
                     }