+++ src/Device/deviceSlice.ts
@@ -8,7 +8,8 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import type { AutoDetectTypes } from '@serialport/bindings-cpp';
 import { SerialPortOpenOptions } from 'serialport';
 
-import { NrfutilDevice, SerialPort } from '../../nrfutil/device/common';
+import { DeviceInfo } from '../../nrfutil';
+import { NrfutilDevice } from '../../nrfutil/device/common';
 import type { RootState } from '../store';
 import {
     getPersistedIsFavorite,
@@ -20,65 +21,74 @@ import {
 } from '../utils/persistentStore';
 
 export interface Device extends NrfutilDevice {
-    serialNumber: string;
-    boardVersion?: string;
     nickname?: string;
-    serialport?: SerialPort;
     favorite?: boolean;
-    id: number;
     persistedSerialPortOptions?: SerialPortOpenOptions<AutoDetectTypes>;
 }
 
+const findDeviceItem = (
+    devices: Device[],
+    id: number,
+    serialNumber?: string
+) => {
+    const index = devices.findIndex(
+        d => d.id === id || d.serialNumber === serialNumber
+    );
+
+    return { index, device: devices[index] };
+};
+
 const updateDevice = (
     state: DeviceState,
-    serialNumber: string,
-    updateToMergeIn: Partial<Device>
+    updateToMergeIn: Partial<Device>,
+    id: number,
+    serialNumber?: string
 ) => {
-    const device = state.devices.get(serialNumber);
+    const device = findDeviceItem(state.devices, id, serialNumber).device;
     if (device) {
         Object.assign(device, updateToMergeIn);
     }
 };
 
 export interface DeviceState {
-    devices: Map<string, Device>;
-    deviceInfo: Device | null;
-    selectedSerialNumber: string | null;
+    devices: Device[];
     readbackProtection: 'unknown' | 'protected' | 'unprotected';
+    selectedDevice?: Device;
+    selectedDeviceInfo?: DeviceInfo;
 }
 
 const initialState: DeviceState = {
-    devices: new Map(),
-    selectedSerialNumber: null,
-    deviceInfo: null,
+    devices: [],
     readbackProtection: 'unknown',
 };
 
 const setPersistedData = (device: Device) => {
-    const newDevice = {
-        ...device,
-        favorite: getPersistedIsFavorite(device.serialNumber),
-        nickname: getPersistedNickname(device.serialNumber),
-    };
+    if (device.serialNumber) {
+        const newDevice = { ...device };
+        newDevice.favorite = getPersistedIsFavorite(device.serialNumber);
+        newDevice.nickname = getPersistedNickname(device.serialNumber);
 
-    const persistedSerialPortSettings = getPersistedSerialPortSettings(
-        newDevice.serialNumber
-    );
+        const persistedSerialPortSettings = getPersistedSerialPortSettings(
+            device.serialNumber
+        );
 
-    if (persistedSerialPortSettings) {
-        const path =
-            newDevice.serialPorts?.[persistedSerialPortSettings.vComIndex]
-                ?.comName;
+        if (persistedSerialPortSettings) {
+            const path =
+                newDevice.serialPorts?.[persistedSerialPortSettings.vComIndex]
+                    ?.comName;
 
-        if (path) {
-            newDevice.persistedSerialPortOptions = {
-                ...persistedSerialPortSettings.serialPortOptions,
-                path,
-            };
+            if (path) {
+                newDevice.persistedSerialPortOptions = {
+                    ...persistedSerialPortSettings.serialPortOptions,
+                    path,
+                };
+            }
         }
+
+        return newDevice;
     }
 
-    return newDevice;
+    return device;
 };
 
 const slice = createSlice({
@@ -89,104 +99,158 @@ const slice = createSlice({
          * Indicates that a device has been selected.
          */
         selectDevice: (state, action: PayloadAction<Device>) => {
-            state.selectedSerialNumber = action.payload.serialNumber;
+            state.selectedDevice = action.payload;
+        },
+
+        setSelectedDeviceInfo: (
+            state,
+            action: PayloadAction<DeviceInfo | undefined>
+        ) => {
+            state.selectedDeviceInfo = action.payload;
         },
 
         /*
          * Indicates that the currently selected device has been deselected.
          */
         deselectDevice: state => {
-            state.selectedSerialNumber = null;
-            state.deviceInfo = null;
+            state.selectedDevice = undefined;
+            state.selectedDeviceInfo = undefined;
             state.readbackProtection = 'unknown';
         },
 
         addDevice: (state, action: PayloadAction<Device>) => {
-            state.devices.set(
-                action.payload.serialNumber,
-                setPersistedData(action.payload)
+            const index = state.devices.findIndex(
+                item =>
+                    item.serialNumber === action.payload.serialNumber ||
+                    item.id === action.payload.id
             );
+
+            const device = setPersistedData(action.payload);
+            if (index !== -1) {
+                state.devices[index] = device;
+            } else {
+                state.devices.push(device);
+            }
         },
 
         persistSerialPortOptions: (
             state,
             action: PayloadAction<SerialPortOpenOptions<AutoDetectTypes>>
         ) => {
-            if (state.selectedSerialNumber === null) return;
+            if (!state.selectedDevice) return;
 
-            const selectedDevice = state.devices.get(
-                state.selectedSerialNumber
-            );
+            const selectedDevice = state.selectedDevice;
 
             if (selectedDevice) {
                 const vComIndex = selectedDevice.serialPorts?.findIndex(
                     e => e.comName === action.payload.path
                 );
 
-                if (vComIndex !== undefined && vComIndex !== -1) {
+                if (
+                    selectedDevice.serialNumber &&
+                    vComIndex !== undefined &&
+                    vComIndex !== -1
+                ) {
                     // eslint-disable-next-line @typescript-eslint/no-unused-vars -- used to filter out the path property
                     const { path: _, ...serialPortOptions } = action.payload;
 
                     persistSerialPortSettingsToStore(
-                        state.selectedSerialNumber,
+                        selectedDevice.serialNumber,
                         {
                             serialPortOptions,
                             vComIndex,
                         }
                     );
 
-                    state.devices.set(state.selectedSerialNumber, {
-                        ...selectedDevice,
-                        persistedSerialPortOptions: action.payload,
-                    });
+                    updateDevice(
+                        state,
+                        {
+                            ...selectedDevice,
+                            persistedSerialPortOptions: action.payload,
+                        },
+                        selectedDevice.id,
+                        selectedDevice.serialNumber
+                    );
                 }
             }
         },
 
         removeDevice: (state, action: PayloadAction<Device>) => {
-            state.devices.delete(action.payload.serialNumber);
+            const index = findDeviceItem(
+                state.devices,
+                action.payload.id,
+                action.payload.serialNumber
+            ).index;
+            state.devices.splice(index, 1);
 
-            if (state.selectedSerialNumber === action.payload.serialNumber) {
-                state.selectedSerialNumber = null;
-                state.deviceInfo = null;
+            if (action.payload.id === state.selectedDevice?.id) {
+                state.selectedDevice = undefined;
+                state.selectedDeviceInfo = undefined;
             }
         },
 
-        toggleDeviceFavorited: (state, action: PayloadAction<string>) => {
-            const newFavoriteState = !state.devices.get(action.payload)
-                ?.favorite;
-            persistIsFavorite(action.payload, newFavoriteState);
-            updateDevice(state, action.payload, {
-                favorite: newFavoriteState,
-            });
+        toggleDeviceFavorited: (state, action: PayloadAction<Device>) => {
+            const device = findDeviceItem(
+                state.devices,
+                action.payload.id,
+                action.payload.serialNumber
+            ).device;
+
+            if (!device.serialNumber) return;
+
+            const newFavoriteState = !device.favorite;
+
+            persistIsFavorite(device.serialNumber, newFavoriteState);
+            updateDevice(
+                state,
+                {
+                    favorite: newFavoriteState,
+                },
+                action.payload.id,
+                action.payload.serialNumber
+            );
         },
 
         setDeviceNickname: {
-            prepare: (serialNumber: string, nickname: string) => ({
-                payload: { serialNumber, nickname },
+            prepare: (device: Device, nickname: string) => ({
+                payload: { device, nickname },
             }),
             reducer: (
                 state,
                 action: PayloadAction<{
-                    serialNumber: string;
+                    device: Device;
                     nickname: string;
                 }>
             ) => {
+                if (!action.payload.device.serialNumber) return;
+
                 persistNickname(
-                    action.payload.serialNumber,
+                    action.payload.device.serialNumber,
                     action.payload.nickname
                 );
-                updateDevice(state, action.payload.serialNumber, {
-                    nickname: action.payload.nickname,
-                });
+                updateDevice(
+                    state,
+                    {
+                        nickname: action.payload.nickname,
+                    },
+                    action.payload.device.id,
+                    action.payload.device.serialNumber
+                );
             },
         },
 
-        resetDeviceNickname: (state, action: PayloadAction<string>) => {
-            persistNickname(action.payload, '');
-            updateDevice(state, action.payload, {
-                nickname: '',
-            });
+        resetDeviceNickname: (state, action: PayloadAction<Device>) => {
+            if (!action.payload.serialNumber) return;
+
+            persistNickname(action.payload.serialNumber, '');
+            updateDevice(
+                state,
+                {
+                    nickname: '',
+                },
+                action.payload.id,
+                action.payload.serialNumber
+            );
         },
 
         setReadbackProtected: (
@@ -204,6 +268,7 @@ export const {
         deselectDevice,
         resetDeviceNickname,
         selectDevice,
+        setSelectedDeviceInfo,
         addDevice,
         removeDevice,
         setDeviceNickname,
@@ -213,22 +278,21 @@ export const {
     },
 } = slice;
 
-export const getDevice = (serialNumber: string) => (state: RootState) =>
-    state.device.devices.get(serialNumber);
+export const getDevice = (state: RootState, serialNumber: string) =>
+    state.device.devices.find(d => d.serialNumber === serialNumber);
 
 export const getDevices = (state: RootState) => state.device.devices;
 
 export const deviceIsSelected = (state: RootState) =>
-    state.device.selectedSerialNumber != null;
+    !!state.device.selectedDevice;
 
-export const selectedDevice = (state: RootState) =>
-    state.device.selectedSerialNumber !== null
-        ? state.device.devices.get(state.device.selectedSerialNumber)
-        : undefined;
+export const selectedDevice = (state: RootState) => state.device.selectedDevice;
 
-export const deviceInfo = (state: RootState) => state.device.deviceInfo;
+export const selectedDeviceInfo = (state: RootState) =>
+    state.device.selectedDevice ? state.device.selectedDeviceInfo : undefined;
+
 export const selectedSerialNumber = (state: RootState) =>
-    state.device.selectedSerialNumber;
+    state.device.selectedDevice?.serialNumber;
 
 export const getReadbackProtection = (state: RootState) =>
     state.device.readbackProtection;